requirements.txt
# Core data analysis libraries
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Graph analysis
networkx>=3.1
python-louvain>=0.16

# NLP
nltk>=3.8.1
spacy>=3.7.0

# Database
duckdb>=0.9.0

# Google Cloud & AI
google-cloud-aiplatform>=1.36.0
vertexai>=0.0.1

# Utilities
jsonschema>=4.19.0
loguru>=0.7.0
autogen[gemini]>=0.2.0

# Additional dependencies
tensorboard>=2.14.0